requirements.txt
fastapi>=0.115.0,<1.0.0
bcrypt>=4.2.0,<5.0.0
httpx>=0.28.0,<1.0.0
PyJWT>=2.9.0,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
pydantic-settings>=2.6.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
uvicorn[standard]>=0.32.0,<1.0.0
